feat: add day 02 solution
Comments: I tried to go for an ad-hoc solution that runs in O(N)
time for each report where N is the length of the report. I intentionally
avoided the naive "remove an element and check again" solution which
runs in O(N^2) but in the end still needed to write it for debugging because
the case analysis got very hairy.